Moved RmcLogoText to CommonUI
